Add an additional upload status ('processed') when we successfully extract marc records to the database

+++ app/jobs/extract_marc_record_metadata_job.rb
@@ -7,18 +7,26 @@ class ExtractMarcRecordMetadataJob < ApplicationJob
 
   # rubocop:disable Metrics/AbcSize
   def perform(upload)
+    return unless upload.active?
+
     progress.total = 0
 
     upload.with_lock do
+      upload.update(status: 'active')
+
       upload.marc_records.delete_all
 
-      upload.each_marc_record_metadata.each_slice(100) do |batch|
+      total = 0
+      upload.read_marc_record_metadata.each_slice(100) do |batch|
+        total += batch.size
         progress.increment(batch.size)
 
         # rubocop:disable Rails/SkipsModelValidations
         MarcRecord.insert_all(batch.map { |x| x.attributes.except('id') }, returning: false)
         # rubocop:enable Rails/SkipsModelValidations
       end
+
+      upload.update(status: 'processed', marc_records_count: total)
     end
 
     UpdateOrganizationStatisticsJob.perform_later(upload.organization, upload.stream, upload)
